chore: rename home modules hook file
